Fix retry logic in task
Apparently, retry needs an `until` attribute to determine when the
task should be considered successful, since it's not smart enough to
figure this out on its own.

Also, while we're at it, a delay is introduced to smooth over any
potential temporary network issues.

Should have been done in c8b149f.