Make it obvious where plugin tasks are running
Running a JCasC deployment can result in Java errors if the Ansible
control node does not have Java installed. This is not immediately clear
from the logs, and it looks like the target machine does not have Java
(or it is somehow not in the PATH). Changing the task name makes it
easier to understand what is running where.