Install plugins directly to a clean plugins dir
This commit changes the way that plugins are installed so that they are
installed directly to the plugins directory (and not /tmp/plugins and
copied), which is also removed before running PIMT. Also, this commit
only runs PIMT if the plugins.yaml file has changed. Also, it is worth
mentioning that at some point, PIMT stopped messing around with
directory permissions, so we no longer need a task to create that
directory.

The advantages of this approach are:

- It avoids potential conflicts with multiple versions of plugins
- It correctly handles updates that remove plugins from plugins.yaml
- It is more efficient and avoids a somewhat lengthy `copy` task (a
  small benefit, but nonetheless a benefit)

Previously, we have just been redeploying our controller whenever we ran
into either of the first two issues, which is a pretty big hammer for
those particular nails.

One disadvantage of this approach is that it is less obvious what
plugins have changed. This will be addressed in the next commit.