Add variable: jenkins_show_plugins_yaml_changes
In the previous commit, we changed the way that plugins are installed to
no longer copy them to a temporary directory first. This makes it a bit
harder to see which plugins have changed when PIMT runs, so this
variable helps to provide that information. To ease the migration from
the previous version, the default value of this variable is true.