Allow PIMT to be re-run in case of failure
AS the comment notes, we can end up in a bad state if PIMT fails for
other reasons (like a network glitch), since we only run PIMT if the
task to copy the plugins.yaml file changed, so re-running this role
would effectively no-op when running PIMT.

Although it would be nicer to use a handler here, they can only be used
for successful tasks, not failed ones. As such, we need to use
block/rescue.