Allow PIMT's output to be logged to file
This tool can generate a lot of output, and extracting and reformatting
this output from an Ansible log is a bit of a chore. This commit
introduces a new variable, jenkins_pimt_logfile, which (when defined)
will create a logfile at the desired location on the Ansible control
machine instead. This variable is undefined (well, technically defined
to the empty string) by default to preserve the current behavior.