Require plugins.yaml to contain the JCasC plugin
It is no longer viable for this role to ship the JCasC plugin itself,
for the following reasons:

- The JCasC plugin has its own dependencies, some of which are shared by
  other Jenkins plugins (namely snakeyaml-api and caffeine-api). We
  don't pin these plugin versions, and they may differ from the versions
  that PIMT would calculate when processing plugins.yaml. In some cases,
  this can result in incompatibilities and some plugins being unable to
  be loaded.
- It prevents users from defining their own JCasC plugin version, which
  may be needed to pin to an older version or to try out new beta
  versions.
- When a new JCasC version is released, it requires this role to be
  bumped and a new release made.

This commit does away with our previous requirement of *not* defining
the JCasC plugin in plugins.yaml, and now *requires* it to be defined
there. :) It also does away with the now-unused Gradle mini-project
which was needed to calculate the JCasC dependencies, making PIMT the
sole plugin dependency calculation tool.

This is of course a breaking change, and thus the minor version will be
bumped in the next commit.